requirements.txt
numpy==1.22.3
opencv_python==4.6.0.66
Pillow==9.3.0
scipy==1.8.1
torch==1.12.1
torchvision==0.13.1
EA-framework==2.2.1